Refactor(login): AppleLoginButton을 UIView에서 UIViewController로 변경
